+++ delivery_carrier_label_postlogistics/data/res_partner.xml
