+++ delivery_carrier_label_postlogistics/i18n/delivery_carrier_label_postlogistics.pot
@@ -206,12 +206,6 @@ msgid "EPS"
 msgstr ""
 
 #. module: delivery_carrier_label_postlogistics
-#: code:addons/delivery_carrier_label_postlogistics/postlogistics/web_service.py:73
-#, python-format
-msgid "Error 401"
-msgstr ""
-
-#. module: delivery_carrier_label_postlogistics
 #: model:delivery.carrier.template.option,name:delivery_carrier_label_postlogistics.postlogistics_layout_option_a5
 msgid "Format A5"
 msgstr ""
