+++ delivery_carrier_label_postlogistics/models/__init__.py
@@ -1,5 +1,8 @@
 # -*- coding: utf-8 -*-
 # © 2013-2015 Yannick Vaucher (Camptocamp SA)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-from . import models
-from . import postlogistics
+from . import company
+from . import res_config
+from . import delivery
+from . import stock
+from . import stock_quant_package
