+++ delivery_carrier_label_postlogistics/views/res_config.xml
